Plugin Directory: Search: Transform Block Directory block searches over to ElasticSearch.
Block Searches might be less reliable today while this change is finalised, a few other index alterations are needed for all current blocks to be properly indexed.
Some changes to the boosting phrases will also be needed most likely once this is finalised and we can do some proper testing.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10117 74240141-8908-4e6f-9713-ba540dce6ec7