Plugin Directory: Filter blocks list to ensure main block is first
The Block Directory API endpoint (on individual WordPress sites) displays the title of the first block in the list. In most cases, there is only one block and this is fine. Other times, there may be child blocks, and in this case we want to make sure the parent block is used as the source of the title.



git-svn-id: https://meta.svn.wordpress.org/sites/trunk@10256 74240141-8908-4e6f-9713-ba540dce6ec7